+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 See respective changelogs for each module.
 
+* [mirrord-protocol](./mirrord-protocol/CHANGELOG.md)
+* [mirrord-agent](./mirrord-agent/CHANGELOG.md)
+* [mirrord-layer](./mirrord-layer/CHANGELOG.md)
 * [VS Code Extension](./vscode-ext/CHANGELOG.md)
-* [CLI](./cli/CHANGELOG.md)
-* [Node Library](./core/CHANGELOG.md)